@@ -4,8 +4,8 @@ ActiveAdmin.register AdminNote do
 
   form do |f|
     f.inputs do
-      f.input :order_id, as: :string, input_html: {readonly: true, value: order.id}
-      f.input :admin_id, input_html: {readonly: true, value: current_user[:id]}
+      f.input :order_id, as: :string, input_html: {readonly: true, value: order.id}, wrapper_html: { style: "display:none" }
+      f.input :admin_id, input_html: {readonly: true, value: current_user[:id]}, wrapper_html: { style: "display:none" }
       f.input :note_type, :as => :select, :collection => AdminNote::TYPES.collect { |type| [type[1].humanize, type[0]] }
       f.input :description
     end
