@@ -52,7 +52,8 @@ class Order < ApplicationRecord
       seller_rejected_other: 'seller_rejected_other'.freeze,
       seller_rejected: 'seller_rejected'.freeze,
       buyer_rejected: 'buyer_rejected'.freeze,
-      buyer_lapsed: 'buyer_lapsed'.freeze
+      buyer_lapsed: 'buyer_lapsed'.freeze,
+      admin_canceled: 'admin_canceled'.freeze
     }
   }.freeze
 
@@ -90,6 +91,7 @@ class Order < ApplicationRecord
   has_many :transactions, dependent: :destroy
   has_many :state_histories, dependent: :destroy
   has_many :admin_notes, dependent: :destroy
+  has_many :fraud_reviews, dependent: :destroy
   has_many :offers, dependent: :destroy
   belongs_to :last_offer, class_name: 'Offer', optional: true
 
