@@ -45,7 +45,7 @@ describe Api::GraphqlController, type: :request do
       Fabricate(:line_item, order: order, list_price_cents: 1000_00, artwork_id: 'a-1', artwork_version_id: '1')
     end
     let(:offer) { Fabricate(:offer, order: order, from_id: seller_id, from_type: 'gallery', amount_cents: 800_00, shipping_total_cents: 100_00, tax_total_cents: 300_00) }
-    let(:artwork) { { _id: 'a-1', current_version_id: '1' } }
+    let(:artwork) { gravity_v1_artwork(_id: 'a-1', current_version_id: '1') }
 
     let(:mutation) do
       <<-GRAPHQL
