@@ -24,7 +24,7 @@ describe OrderFollowUpJob, type: :job do
         context 'Buy order' do
           it 'transitions a submitted order to seller_lapsed' do
             Timecop.freeze(order.state_expires_at + 1.second) do
-              expect_any_instance_of(OrderCancellationService).to receive(:seller_lapse!)
+              expect(OrderService).to receive(:seller_lapse!)
               OrderFollowUpJob.perform_now(order.id, Order::SUBMITTED)
             end
           end
@@ -38,7 +38,7 @@ describe OrderFollowUpJob, type: :job do
           context 'Last offer from seller (awaiting response from buyer)' do
             it 'transitions a submitted order to buyer_lapsed' do
               Timecop.freeze(order.state_expires_at + 1.second) do
-                expect_any_instance_of(OrderCancellationService).to receive(:buyer_lapse!)
+                expect(OrderService).to receive(:buyer_lapse!)
                 OrderFollowUpJob.perform_now(order.id, Order::SUBMITTED)
               end
             end
@@ -47,7 +47,7 @@ describe OrderFollowUpJob, type: :job do
             let(:offer) { Fabricate(:offer, from_id: buyer_id, order: order, from_type: buyer_type, submitted_at: Time.now.utc) }
             it 'transitions a submitted order to seller_lapsed' do
               Timecop.freeze(order.state_expires_at + 1.second) do
-                expect_any_instance_of(OrderCancellationService).to receive(:seller_lapse!)
+                expect(OrderService).to receive(:seller_lapse!)
                 OrderFollowUpJob.perform_now(order.id, Order::SUBMITTED)
               end
             end
@@ -69,7 +69,7 @@ describe OrderFollowUpJob, type: :job do
         order.update!(state: Order::SUBMITTED)
         Timecop.freeze(order.state_expires_at + 1.second) do
           expect(OrderService).to_not receive(:abandon!)
-          expect_any_instance_of(OrderCancellationService).to_not receive(:seller_lapse!)
+          expect(OrderService).to_not receive(:seller_lapse!)
           OrderFollowUpJob.perform_now(order.id, Order::PENDING)
         end
       end
@@ -77,7 +77,7 @@ describe OrderFollowUpJob, type: :job do
     context 'with an order in the same state before its expiration time' do
       it 'does nothing' do
         expect(OrderService).to_not receive(:abandon!)
-        expect_any_instance_of(OrderCancellationService).to_not receive(:seller_lapse!)
+        expect(OrderService).to_not receive(:seller_lapse!)
         OrderFollowUpJob.perform_now(order.id, Order::PENDING)
       end
     end
