@@ -1,6 +1,7 @@
-if ENV.fetch('CI', false)
-  require 'coveralls'
-  Coveralls.wear!('rails')
+require 'simplecov'
+
+SimpleCov.start do
+  add_filter '/spec/'
 end
 
 RSpec.configure do |config|
